--- a/src/core/file/fileCollect.ts
+++ b/src/core/file/fileCollect.ts
@@ -1,22 +1,12 @@
 import pc from 'picocolors';
-import { Piscina } from 'piscina';
 import { logger } from '../../shared/logger.js';
-import { getWorkerThreadCount } from '../../shared/processConcurrency.js';
+import { initPiscina } from '../../shared/processConcurrency.js';
 import type { RepomixProgressCallback } from '../../shared/types.js';
 import type { RawFile } from './fileTypes.js';
 import type { FileCollectTask } from './workers/fileCollectWorker.js';
 
 const initTaskRunner = (numOfTasks: number) => {
-  const { minThreads, maxThreads } = getWorkerThreadCount(numOfTasks);
-  logger.trace(`Initializing worker pool with min=${minThreads}, max=${maxThreads} threads`);
-
-  const pool = new Piscina({
-    filename: new URL('./workers/fileCollectWorker.js', import.meta.url).href,
-    minThreads,
-    maxThreads,
-    idleTimeout: 5000,
-  });
-
+  const pool = initPiscina(numOfTasks, new URL('./workers/fileCollectWorker.js', import.meta.url).href);
   return (task: FileCollectTask) => pool.run(task);
 };