--- a/src/core/metrics/aggregateMetrics.ts
+++ b/src/core/metrics/aggregateMetrics.ts
@@ -1,16 +1,20 @@
+import { TiktokenEncoding } from 'tiktoken';
 import type { ProcessedFile } from '../file/fileTypes.js';
-import type { TokenCounter } from '../tokenCount/tokenCount.js';
+import { TokenCounter } from '../tokenCount/tokenCount.js';
 import type { FileMetrics } from './calculateIndividualFileMetrics.js';
 
 export const aggregateMetrics = (
   fileMetrics: FileMetrics[],
   processedFiles: ProcessedFile[],
   output: string,
-  tokenCounter: TokenCounter,
+  tokenCounterEncoding: TiktokenEncoding,
 ) => {
   const totalFiles = processedFiles.length;
   const totalCharacters = output.length;
+  const tokenCounter = new TokenCounter(tokenCounterEncoding);
   const totalTokens = tokenCounter.countTokens(output);
+
+  tokenCounter.free();
 
   const fileCharCounts: Record<string, number> = {};
   const fileTokenCounts: Record<string, number> = {};