… package.json
… src/core/metrics/calculateIndividualFileMetrics.ts
… src/core/metrics/calculateOutputMetrics.ts
… src/core/metrics/workers/fileMetricsWorker.ts
… src/core/metrics/workers/metricsWorker.ts
… src/core/metrics/workers/outputMetricsWorker.ts
… src/core/metrics/workers/types.ts
… src/core/packager.ts
… src/core/security/runSecurityCheckIfEnabled.ts
… src/core/security/securityCheck.ts
… src/core/security/validateFileSafety.ts
… src/core/security/workers/securityCheckWorker.ts
… src/core/tokenCount/tokenCount.ts
… src/shared/processConcurrency.ts
… tests/cli/actions/remoteAction.test.ts
… tests/cli/cliPrint.test.ts
… tests/config/configSchema.test.ts
… tests/core/file/fileCollect.test.ts
… tests/core/file/fileProcess.test.ts
… tests/core/metrics/aggregateMetrics.test.ts
… tests/core/metrics/calculateAllFileMetrics.test.ts
… tests/core/metrics/calculateIndividualFileMetrics.test.ts
… tests/core/metrics/calculateMetrics.test.ts
… tests/core/metrics/calculateOutputMetrics.test.ts
… tests/core/packager.test.ts
… tests/core/packager/copyToClipboardIfEnabled.test.ts
… tests/core/security/runSecurityCheckIfEnabled.test.ts
… tests/core/security/securityCheck.test.ts
… tests/core/security/validateFileSafety.test.ts
… tests/core/security/workers/securityCheckWorker.test.ts